feat: added the option to ignore keys that are not defined in the keymap
